+++ src/main/java/online/kingdomkeys/kingdomkeys/client/ClientSetup.java
@@ -5,16 +5,19 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.PlayerModel;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.client.renderer.ShaderInstance;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.player.PlayerRenderer;
 import net.minecraft.client.resources.PlayerSkin;
+import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -30,11 +33,9 @@ import net.neoforged.neoforge.client.gui.VanillaGuiLayers;
 import net.neoforged.neoforge.common.NeoForge;
 import online.kingdomkeys.kingdomkeys.KingdomKeys;
 import online.kingdomkeys.kingdomkeys.client.gui.overlay.*;
-import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
-import net.minecraft.client.resources.model.BakedModel;
-import online.kingdomkeys.kingdomkeys.client.render.*;
 import online.kingdomkeys.kingdomkeys.client.particles.KeybladeHitParticle;
 import online.kingdomkeys.kingdomkeys.client.particles.ModParticles;
+import online.kingdomkeys.kingdomkeys.client.render.*;
 import online.kingdomkeys.kingdomkeys.client.render.item.KeychainModelWrapper;
 import online.kingdomkeys.kingdomkeys.client.render.item.KeychainRenderer;
 import online.kingdomkeys.kingdomkeys.config.ModConfigs;
@@ -45,6 +46,7 @@ import online.kingdomkeys.kingdomkeys.entity.ModEntities;
 import online.kingdomkeys.kingdomkeys.handler.ClientEvents;
 import online.kingdomkeys.kingdomkeys.handler.InputHandler;
 import online.kingdomkeys.kingdomkeys.item.KeybladeItem;
+import online.kingdomkeys.kingdomkeys.item.ModItems;
 import online.kingdomkeys.kingdomkeys.util.Utils;
 
 import java.io.IOException;
@@ -53,84 +55,69 @@ import java.util.Map.Entry;
 @EventBusSubscriber(value = Dist.CLIENT)
 public class ClientSetup {
 
-	public static ResourceLocation
-			COMMAND_MENU,
-			PLAYER_PORTRAIT,
-			PLAYER_PORTRAIT_CROWN,
-			HP_BAR,
-			MP_BAR,
-			DRIVE_BAR,
-			KK_NOTIFICATIONS,
-			LOCK_ON,
-			PARTY_INFO,
-			SHOTLOCK,
-			STATION_OF_AWAKENING_MESSAGES;
+	public static ResourceLocation COMMAND_MENU, PLAYER_PORTRAIT, PLAYER_PORTRAIT_CROWN, HP_BAR, MP_BAR, DRIVE_BAR, KK_NOTIFICATIONS, LOCK_ON, PARTY_INFO, SHOTLOCK, STATION_OF_AWAKENING_MESSAGES;
+	public static ShaderInstance hpShader, focusShader, shotlockShader, gummiHPShader;
+	private static KeychainRenderer keychainRenderer;
 
-    //Register the entity models
-    @SubscribeEvent
-    public static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
-        ModEntities.registerRenderers(event);
-    }
+	//Register the entity models
+	@SubscribeEvent
+	public static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
+		ModEntities.registerRenderers(event);
+	}
 
 	@SubscribeEvent
-	public static void registerLayers(
-			EntityRenderersEvent.RegisterLayerDefinitions event
-	) {
+	public static void registerLayers(EntityRenderersEvent.RegisterLayerDefinitions event) {
 		ModEntities.registerLayers(event);
 
-		event.registerLayerDefinition(
-				OrganizationArmorOverlayRenderer.OUTER_LAYER,
-				OrganizationArmorOverlayRenderer::createOuterLayer
-		);
+		event.registerLayerDefinition(ClothArmorOverlayRenderer.BASE_OUTER_LAYER, ClothArmorOverlayRenderer::createBaseOuterLayer);
+		event.registerLayerDefinition(ClothArmorOverlayRenderer.BASE_LEGGINGS_LAYER, ClothArmorOverlayRenderer::createBaseLeggingsLayer);
 
-		event.registerLayerDefinition(
-				OrganizationArmorOverlayRenderer.LEGGINGS_LAYER,
-				OrganizationArmorOverlayRenderer::createLeggingsLayer
-		);
+		event.registerLayerDefinition(ClothArmorOverlayRenderer.OUTER_LAYER, ClothArmorOverlayRenderer::createOuterLayer);
+		event.registerLayerDefinition(ClothArmorOverlayRenderer.LEGGINGS_LAYER, ClothArmorOverlayRenderer::createLeggingsLayer);
 	}
 
 	@SubscribeEvent
 	public static void addLayers(EntityRenderersEvent.AddLayers event) {
-		for(Entry<EntityType<?>, EntityRenderer<?>> entry : Minecraft.getInstance().getEntityRenderDispatcher().renderers.entrySet()) {
-			if(entry.getValue() instanceof LivingEntityRenderer renderer && !(entry.getValue() instanceof PlayerRenderer)) {
+		for (Entry<EntityType<?>, EntityRenderer<?>> entry : Minecraft.getInstance().getEntityRenderDispatcher().renderers.entrySet()) {
+			if (entry.getValue() instanceof LivingEntityRenderer renderer && !(entry.getValue() instanceof PlayerRenderer)) {
 				renderer.addLayer(new AeroLayerRenderer<>(renderer, event.getEntityModels()));
 				renderer.addLayer(new CrownLayerRenderer<>(renderer, event.getEntityModels()));
 				renderer.addLayer(new FreezeLayerRenderer<>(renderer, event.getEntityModels()));
 				renderer.addLayer(new KeybladeArmorRenderer<>(renderer, event.getEntityModels()));
-				renderer.addLayer(new OrganizationArmorOverlayRenderer<>(renderer, event.getEntityModels()));
+				renderer.addLayer(new ClothArmorOverlayRenderer<>(renderer, event.getEntityModels()));
 			}
 		}
 
 		LivingEntityRenderer<Player, PlayerModel<Player>> renderer = event.getSkin(PlayerSkin.Model.WIDE);
 		renderer.addLayer(new DriveLayerRenderer<>(renderer));
 		renderer.addLayer(new StopLayerRenderer<>(renderer, event.getEntityModels()));
-		renderer.addLayer(new ShoulderLayerRenderer<>(renderer, event.getEntityModels(),true));
+		renderer.addLayer(new ShoulderLayerRenderer<>(renderer, event.getEntityModels(), true));
 		renderer.addLayer(new KeybladeArmorRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new AeroLayerRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new CrownLayerRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new FreezeLayerRenderer<>(renderer, event.getEntityModels()));
-		renderer.addLayer(new OrganizationArmorOverlayRenderer<>(renderer, event.getEntityModels()));
+		renderer.addLayer(new ClothArmorOverlayRenderer<>(renderer, event.getEntityModels()));
 
 		renderer = event.getSkin(PlayerSkin.Model.SLIM);
 		renderer.addLayer(new DriveLayerRenderer<>(renderer));
 		renderer.addLayer(new StopLayerRenderer<>(renderer, event.getEntityModels()));
-		renderer.addLayer(new ShoulderLayerRenderer<>(renderer, event.getEntityModels(),false));
+		renderer.addLayer(new ShoulderLayerRenderer<>(renderer, event.getEntityModels(), false));
 		renderer.addLayer(new KeybladeArmorRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new AeroLayerRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new CrownLayerRenderer<>(renderer, event.getEntityModels()));
 		renderer.addLayer(new FreezeLayerRenderer<>(renderer, event.getEntityModels()));
-		renderer.addLayer(new OrganizationArmorOverlayRenderer<>(renderer, event.getEntityModels()));
+		renderer.addLayer(new ClothArmorOverlayRenderer<>(renderer, event.getEntityModels()));
 	}
 
 	@SubscribeEvent
 	public static void registerOverlays(RegisterGuiLayersEvent event) {
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("shortcuts"), ShortcutsGui.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("command_menu"), CommandMenuGui.INSTANCE);
-        event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("player_portrait"), PlayerPortraitGui.INSTANCE);
+		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("player_portrait"), PlayerPortraitGui.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("hp_bar"), HPGui.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("player_portrait_crown"), PlayerPortraitGui.CROWN_OVERLAY);
-        event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("gummi_hud"), GummiHUD.INSTANCE);
-        event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("mp_bar"), MPGui.INSTANCE);
+		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("gummi_hud"), GummiHUD.INSTANCE);
+		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("mp_bar"), MPGui.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("drive_bar"), DriveGui.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("kk_notifications"), GuiOverlay.INSTANCE);
 		event.registerBelow(VanillaGuiLayers.CROSSHAIR, KingdomKeys.rl("lock_on"), LockOnGui.INSTANCE);
@@ -142,8 +129,6 @@ public class ClientSetup {
 		event.registerBelow(VanillaGuiLayers.CHAT, KingdomKeys.rl("castle_oblivion_minimap"), COMinimap.INSTANCE);
 		event.registerAbove(VanillaGuiLayers.CHAT, KingdomKeys.rl("item_get"), ItemGetGui.INSTANCE);
 	}
-
-	private static KeychainRenderer keychainRenderer;
 
 	@SubscribeEvent
 	public static void registerClientExtensions(RegisterClientExtensionsEvent event) {
@@ -171,6 +156,26 @@ public class ClientSetup {
 		};
 
 		BuiltInRegistries.ITEM.stream().filter(KeybladeItem.class::isInstance).forEach(keyblade -> event.registerItem(guarding, keyblade));
+
+		IClientItemExtensions clothArmor = new IClientItemExtensions() {
+			@Override
+			public HumanoidModel<?> getHumanoidArmorModel(LivingEntity entity, ItemStack stack, EquipmentSlot slot, HumanoidModel<?> original) {
+				return ClothArmorOverlayRenderer.baseModel(slot);
+			}
+		};
+
+		event.registerItem(clothArmor,
+				ModItems.organizationRobe_Chestplate.get(), ModItems.organizationRobe_Leggings.get(), ModItems.organizationRobe_Boots.get(),
+				ModItems.xemnas_Chestplate.get(), ModItems.xemnas_Leggings.get(), ModItems.xemnas_Boots.get(),
+				ModItems.antiCoat_Chestplate.get(), ModItems.antiCoat_Leggings.get(), ModItems.antiCoat_Boots.get(),
+				ModItems.ira_Chestplate.get(), ModItems.ira_Leggings.get(), ModItems.ira_Boots.get(),
+				ModItems.invi_Chestplate.get(), ModItems.invi_Leggings.get(), ModItems.invi_Boots.get(),
+				ModItems.aced_Chestplate.get(), ModItems.aced_Leggings.get(), ModItems.aced_Boots.get(),
+				ModItems.gula_Chestplate.get(), ModItems.gula_Leggings.get(), ModItems.gula_Boots.get(),
+				ModItems.ava_Chestplate.get(), ModItems.ava_Leggings.get(), ModItems.ava_Boots.get(),
+				ModItems.dark_Riku_Chestplate.get(), ModItems.dark_Riku_Leggings.get(), ModItems.dark_Riku_Boots.get(),
+				ModItems.vanitas_Chestplate.get(), ModItems.vanitas_Leggings.get(), ModItems.vanitas_Boots.get(),
+				ModItems.vanitas_Remnant_Chestplate.get(), ModItems.vanitas_Remnant_Leggings.get(), ModItems.vanitas_Remnant_Boots.get());
 	}
 
 	@SubscribeEvent
@@ -190,24 +195,23 @@ public class ClientSetup {
 		ResourceLocation o = event.getName();
 		PlayerData playerData = PlayerData.get(player);
 		GlobalData globalData = GlobalData.get(player);
-		if(playerData == null || globalData == null)
-			return;
-		
-		if(!Utils.shouldRenderOverlay(player)) { //If it shouldn't render cause it's set to HIDE or WEAPON and not holding one
+		if (playerData == null || globalData == null) return;
+
+		if (!Utils.shouldRenderOverlay(player)) { //If it shouldn't render cause it's set to HIDE or WEAPON and not holding one
 			event.setCanceled(o.equals(COMMAND_MENU) || o.equals(MP_BAR) || o.equals(DRIVE_BAR) || o.equals(SHOTLOCK)); //Remove all these 4 bars
-			if(o.equals(HP_BAR) || o.equals(PLAYER_PORTRAIT) || o.equals(PLAYER_PORTRAIT_CROWN)) { //Allow HP to be shown if KO'd
+			if (o.equals(HP_BAR) || o.equals(PLAYER_PORTRAIT) || o.equals(PLAYER_PORTRAIT_CROWN)) { //Allow HP to be shown if KO'd
 				event.setCanceled(!player.hasEffect(ModMobEffects.KO));
 			}
 		} else { //If mode is set to SHOW or WEAPON while holding one
-			if(o.equals(MP_BAR)) { //Remove MP Bar is magics map is empty
+			if (o.equals(MP_BAR)) { //Remove MP Bar is magics map is empty
 				event.setCanceled(playerData.getEquippedMagics().isEmpty());
 				return;
 			}
-			if(o.equals(SHOTLOCK)) {
+			if (o.equals(SHOTLOCK)) {
 				event.setCanceled(playerData.getEquippedShotlock().isEmpty());
 				return;
 			}
-			if(o.equals(DRIVE_BAR)) {
+			if (o.equals(DRIVE_BAR)) {
 				event.setCanceled(Utils.getVisibleDriveForms(player).size() <= 1);
 				return;
 			}
@@ -218,9 +222,9 @@ public class ClientSetup {
 		}
 	}
 
-    @OnlyIn(Dist.CLIENT)
-    @SubscribeEvent
-    public static void setupClient(FMLClientSetupEvent event) {
+	@OnlyIn(Dist.CLIENT)
+	@SubscribeEvent
+	public static void setupClient(FMLClientSetupEvent event) {
 		COMMAND_MENU = KingdomKeys.rl("command_menu");
 		PLAYER_PORTRAIT = KingdomKeys.rl("player_portrait");
 		PLAYER_PORTRAIT_CROWN = KingdomKeys.rl("player_portrait_crown");
@@ -241,8 +245,8 @@ public class ClientSetup {
 		NeoForge.EVENT_BUS.register(SoAMessages.INSTANCE);
 		NeoForge.EVENT_BUS.register(ShotlockMinigameGui.INSTANCE);
 		NeoForge.EVENT_BUS.register(new WorldMapRenderer());
-		
-    }
+
+	}
 
 	@OnlyIn(Dist.CLIENT)
 	@SubscribeEvent
@@ -256,6 +260,7 @@ public class ClientSetup {
 	@SubscribeEvent
 	public static void wrapKeybladeModels(ModelEvent.ModifyBakingResult event) {
 		KeychainRenderer.clearCache();
+		ClothArmorOverlayRenderer.clearCache();
 
 		for (ResourceLocation keyblade : KeychainRenderer.splitKeyblades()) {
 			ModelResourceLocation location = ModelResourceLocation.inventory(keyblade);
@@ -272,23 +277,21 @@ public class ClientSetup {
 		}
 	}
 
-	public static ShaderInstance hpShader, focusShader, shotlockShader, gummiHPShader;
-
 	@SubscribeEvent
 	public static void registerShaders(RegisterShadersEvent event) {
 		try {
 			event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("hp"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
-                hpShader = shaderInstance;
+				hpShader = shaderInstance;
 			});
-            event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("focus"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
-                focusShader = shaderInstance;
-            });
-            event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("shotlock"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
-                shotlockShader = shaderInstance;
-            });
-            event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("gummi_hp"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
-                gummiHPShader = shaderInstance;
-            });
+			event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("focus"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
+				focusShader = shaderInstance;
+			});
+			event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("shotlock"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
+				shotlockShader = shaderInstance;
+			});
+			event.registerShader(new ShaderInstance(event.getResourceProvider(), KingdomKeys.rl("gummi_hp"), DefaultVertexFormat.POSITION_TEX), shaderInstance -> {
+				gummiHPShader = shaderInstance;
+			});
 		} catch (IOException e) {
 			e.printStackTrace();
 			throw new RuntimeException("Could not load shader");