+++ src/main/java/online/kingdomkeys/kingdomkeys/item/BaseArmorItem.java
@@ -25,53 +25,39 @@ import java.util.UUID;
 public class BaseArmorItem extends ArmorItem implements IItemCategory {
 
 	String textureName;
-	
+
 	public BaseArmorItem(Holder<ArmorMaterial> materialIn, Type slot, String textureName) {
 		super(materialIn, slot, new Item.Properties().stacksTo(1));
 		this.textureName = textureName;
 	}
 
+	public String getTextureName() {
+		return this.textureName;
+	}
+
 	@Override
-	public ResourceLocation getArmorTexture(
-			ItemStack stack,
-			Entity entity,
-			EquipmentSlot slot,
-			ArmorMaterial.Layer layer,
-			boolean innerModel
-	) {
+	public ResourceLocation getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, ArmorMaterial.Layer layer, boolean innerModel) {
 		if (this.textureName != null) {
 			if (slot == EquipmentSlot.LEGS) {
-				return KingdomKeys.rl(
-						"textures/models/armor/" +
-								this.textureName +
-								"_layer_2.png"
-				);
-			} else if (
-					slot == EquipmentSlot.HEAD ||
-							slot == EquipmentSlot.CHEST ||
-							slot == EquipmentSlot.FEET
-			) {
-				return KingdomKeys.rl(
-						"textures/models/armor/" +
-								this.textureName +
-								"_layer_1.png"
-				);
+				return KingdomKeys.rl("textures/models/armor/" + this.textureName + "_layer_2.png");
+			} else if (slot == EquipmentSlot.HEAD || slot == EquipmentSlot.CHEST || slot == EquipmentSlot.FEET) {
+				return KingdomKeys.rl("textures/models/armor/" + this.textureName + "_layer_1.png");
 			}
 		}
 
 		return null;
 	}
-	
+
 	@OnlyIn(Dist.CLIENT)
 	@Override
 	public void appendHoverText(ItemStack stack, TooltipContext tooltipContext, List<Component> tooltip, TooltipFlag flagIn) {
 		if (flagIn.isAdvanced()) {
-			if(Utils.hasArmorID(stack)) {
+			if (Utils.hasArmorID(stack)) {
 				tooltip.add(Component.translatable(ChatFormatting.RED + "DEBUG:"));
 				tooltip.add(Component.translatable(ChatFormatting.WHITE + Utils.getArmorID(stack).toString()));
 			}
 		}
-    }
+	}
 
 	@Override
 	public void inventoryTick(ItemStack stack, Level worldIn, Entity entityIn, int itemSlot, boolean isSelected) {
@@ -95,7 +81,7 @@ public class BaseArmorItem extends ArmorItem implements IItemCategory {
 		@SubscribeEvent
 		public static void onItemDropped(EntityJoinLevelEvent event) {
 			if (event.getEntity() instanceof ItemEntity) {
-				ItemStack droppedItem = ((ItemEntity)event.getEntity()).getItem();
+				ItemStack droppedItem = ((ItemEntity) event.getEntity()).getItem();
 				UUID droppedID = Utils.getArmorID(droppedItem);
 				if (droppedID != null && droppedItem.getItem() instanceof BaseArmorItem) {
 					event.setCanceled(true);