+++ src/main/java/online/kingdomkeys/kingdomkeys/mixin/epicfight/FirstPersonRendererMixin.java
@@ -4,11 +4,11 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.world.entity.EntityType;
 import online.kingdomkeys.kingdomkeys.client.render.DriveLayerRenderer;
 import online.kingdomkeys.kingdomkeys.client.render.KeybladeArmorRenderer;
-import online.kingdomkeys.kingdomkeys.client.render.OrganizationArmorOverlayRenderer;
+import online.kingdomkeys.kingdomkeys.client.render.ClothArmorOverlayRenderer;
 import online.kingdomkeys.kingdomkeys.client.render.ShoulderLayerRenderer;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.PatchedArmourLayerRenderer;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.PatchedDriveLayerRenderer;
-import online.kingdomkeys.kingdomkeys.integration.epicfight.PatchedOrganizationArmorOverlayRenderer;
+import online.kingdomkeys.kingdomkeys.integration.epicfight.PatchedClothArmorOverlayRenderer;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.PatchedShoulderLayerRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -26,7 +26,7 @@ public class FirstPersonRendererMixin {
         thisOne.addPatchedLayer(DriveLayerRenderer.class, new PatchedDriveLayerRenderer<>(true));
         thisOne.addPatchedLayer(KeybladeArmorRenderer.class, new PatchedArmourLayerRenderer<>(true));
         thisOne.addPatchedLayer(ShoulderLayerRenderer.class, new PatchedShoulderLayerRenderer<>());
-        thisOne.addPatchedLayer(OrganizationArmorOverlayRenderer.class, new PatchedOrganizationArmorOverlayRenderer<>(true));
+        thisOne.addPatchedLayer(ClothArmorOverlayRenderer.class, new PatchedClothArmorOverlayRenderer<>(true));
     }
 
 }